requirements.txt
streamlit
streamlit-drawable-canvas==0.9.0
opencv-python
pillow